+++ src/main/java/com/iybook/sales/controller/CancelManagementController.java
@@ -1,0 +1,13 @@
+package com.iybook.sales.controller;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Slf4j
+@RequiredArgsConstructor
+@RequestMapping("/sales")
+@Controller
+public class CancelManagementController {
+}
